[*] Better tests and rename datasource/local/null -> null_table
