+++ src/error_test.rs
@@ -1,3 +1,7 @@
+// Copyright 2020-2021 The FuseQuery Authors.
+//
+// SPDX-License-Identifier: Apache-2.0.
+
 #[test]
 fn test_build_errors() {
     use crate::error::*;
